+++ resdk/resources/relation.py
@@ -25,8 +25,12 @@ class Relation(BaseResolweResource):
 
     endpoint = 'relation'
 
-    WRITABLE_FIELDS = ('collection', 'label') + BaseResolweResource.WRITABLE_FIELDS
-    UPDATE_PROTECTED_FIELDS = ('entities', 'type') + BaseResolweResource.UPDATE_PROTECTED_FIELDS
+    UPDATE_PROTECTED_FIELDS = BaseResolweResource.UPDATE_PROTECTED_FIELDS + (
+        'entities', 'type',
+    )
+    WRITABLE_FIELDS = BaseResolweResource.WRITABLE_FIELDS + (
+        'collection', 'label',
+    )
 
     ALL_PERMISSIONS = ['view', 'edit', 'share', 'owner']